--- src/components/Window.tsx
@@ -1,7 +1,0 @@
-export function Window() {
-  return (
-    <div className="mockup-window bg-base-300 w-[500px] my-0 mx-auto">
-      <div className="flex justify-center px-4 py-16 bg-base-200">Hello!</div>
-    </div>
-  );
-}
